+++ e2e/auth-flow.spec.ts
@@ -1,5 +1,3 @@
-// `zzz-` prefix forces these tests to run near the end (workers=1, alphabetical order),
-// so login/logout specs do not invalidate the shared storageState used by earlier specs.
 import { expect, test } from "@playwright/test";
 
 test.use({ storageState: { cookies: [], origins: [] } });
